navbar with layout
