--- frontend/src/components/navbar/navbar.css
@@ -1,42 +1,0 @@
-.burger-menu {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    width: 30px;
-    height: 25px;
-    cursor: pointer;
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    z-index: 1000;
-  }
-  
-  .burger-line {
-    width: 100%;
-    height: 3px;
-    background-color: #000;
-  }
-  
-  /* Navbar Overlay Styles */
-  .navbar-overlay {
-    position: fixed;
-    top: 0;
-    left: -100%;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.9);
-    transition: left 0.3s ease-in-out;
-    z-index: 999;
-  }
-  
-  .navbar-overlay.open {
-    left: 0;
-  }
-  
-  .navbar-overlay .AdminNav {
-    position: relative;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    overflow-y: auto;
-  }
